requirements.txt
faiss-cpu
transformers
Pillow
pypdf
jq
anthropic
langchain
#langchain-community==0.0.31
langchain_community
langchain_core
streamlit
botocore
boto3
awscli
requests
requests-aws4auth
opensearch-py
python-dotenv==1.0.0